ai/requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.85.0
h11==0.14.0
idna==3.4
numpy==1.23.3
openai==0.23.1
openpyxl==3.0.10
pandas==1.5.0
pandas-stubs==1.5.0.220926
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tqdm==4.64.1
types-pytz==2022.2.1.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
